fix: keys generation and backend start  (#37)
## fixes KILTProtocol/NO Ticket

The newest implementations on this project were not letting me start the
backend.

The __key generation__ used for the `well-known-did-config` and the
`.env`-variables (and consequently the ones on chain) was different than
the __key generation__ on the backend and therefore the one used for
signing. The `config` verifies this and was throwing.

The `backend-prod` was being concurrently started with the `frontend`
and the `backend` and therefore stealing the backend's port and not
letting him exist.

These troubles should be fixed now.